[MPLUGIN-372] IT
Adding separate IT that demonstrates how this IT FAILS (as expected)
on master, while passes with PR.

The IT itself is based on `annotation-with-inheritance-from-deps`
with following changes:
* dependency maven-plugin-api scope moved to `provided`
* FirstMojo got `@deprecated` javadoc tag REMOVED (to make maven-plugin-plugin search superclasses)

The IT against master shows that build FAILS (so install goal fail) as
maven-plugin-plugin:descriptor goal fails. It reports "API incompatible"
but actual cause is CNFEx.