[MPLUGIN-394] Respect input encoding
PluginReport didn't specify an encoding while reading the
Plugin-Descriptor. This causes the fallback to the plattform encoding
and ignores the (user specified) input encoding.

This commits adds an integration test to reproduce the error and fixes
it by using the input encoding for the creation of the used Reader.

The integration tests forces the plattform encoding to CP1252 (Western
European charset used in Windows) while the source (input) and output
encoding is UTF-8.
The source file contains non-ASCII characters which have a different
representation in both encodings.